backend/requirements.txt

Django==3.2

djangorestframework==3.12.4

djoser==2.2.0

python-decouple==3.8

Pillow==9.0.0